// src/shared/types.d.ts
export interface FileRequest {
    type: 'file_request';
    fileHash: string;
    range: {
        start: number;
        end: number;
    };
}